search-engine/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pymupdf==1.24.0
pandas==2.2.1
impact-factor==1.1.0
pymongo==4.6.3
pydantic-settings==2.0.3
llama-index==0.10.25
llama-index-embeddings-instructor==0.1.3
sentence-transformers==2.2.2
faiss-cpu==1.8.0
llama-index-vector-stores-faiss==0.1.2
openai==1.30.2
llama-index-llms-openai==0.1.20
gradio==4.19.2
pymongo
psycopg2-binary
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.24
